+++ contracts/license/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Changed 'get_licenses' to use feeder for passing return values [#1054]
 - Changed 'use_license' to check if license already nullified [#1051]
 - Changed 'get_licenses' to return values by adding 'pos' to every license returned [#1040]
 - Changed 'issue_license' by removing the 'pos' argument and self position determination [#1039]
@@ -23,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `license` contract to Rusk [#960]
 
+[#1054]: https://github.com/dusk-network/rusk/issues/1054
 [#1052]: https://github.com/dusk-network/rusk/issues/1052
 [#1051]: https://github.com/dusk-network/rusk/issues/1051
 [#1040]: https://github.com/dusk-network/rusk/issues/1040